Move loadLargePayload() to separate file to avoid issues with jest in benchmarks
It complains as follows, which we can just avoid:

file:///project/dist/tests/helpers.js:1
        import { expect } from '@jest/globals';
                 ^^^^^^
        SyntaxError: Named export 'expect' not found. The requested module '@jest/globals' is a CommonJS module, which may not support all module.exports as named exports.
        CommonJS modules can always be imported via the default export, for example using:
